fix: Merge Okta registration tests
